+++ src/gromacs/fileio/h5md_io.h
@@ -44,7 +44,7 @@
 
 struct gmx_mtop_t;
 typedef int64_t hid_t;
-typedef uint64_t hsize_t;
+typedef unsigned long long hsize_t;
 
 /*! \brief The container of the H5MD data. The class is designed to read/write data according to de Buyl et al., 2014
  * (https://www.sciencedirect.com/science/article/pii/S0010465514000447) and https://www.nongnu.org/h5md/h5md.html
