+++ src/gromacs/fileio/h5md_util.h
@@ -42,7 +42,7 @@
 
 
 typedef int64_t hid_t;
-typedef uint64_t hsize_t;
+typedef unsigned long long hsize_t;
 
 /*! \brief An enumeration of compression options */
 enum class CompressionAlgorithm
